requirements.txt
streamlit
requests
beautifulsoup4
trafilatura
google-generativeai
python-dotenv
cloudscraper 
fake-useragent
markdown
lxml-html-clean